backend/requirements.txt
# Core framework
fastapi==0.115.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# Cache / Queue
redis==5.2.0
celery==5.4.0
celery[redis]==5.4.0

# Crawling — httpx pinned to satisfy ollama<0.28 constraint
httpx==0.27.2
playwright==1.49.0
curl_cffi==0.7.0

# HTML parsing & extraction
beautifulsoup4==4.12.3
lxml==5.3.0
html2text==2024.2.26
markdownify==0.14.1
extruct==0.17.0

# ML / NLP
numpy==1.26.4
scikit-learn==1.5.2
scipy==1.14.1
joblib==1.4.2

# LLM
ollama==0.4.0
instructor==1.7.0
openai==1.57.0

# Utilities
python-multipart==0.0.17
python-dotenv==1.0.1
structlog==24.4.0
tenacity==9.0.0
geopy==2.4.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
